Now shows the progress based on the file chunks
Originally, this was sending the total files and the files processed.
The problem we might have with this is that if there is a very large
file, the progress indicator on the page won't move for a while. With
the percentage progress, this is updated each time a slice is calculated
and because the slices are all the same size, there should be reasonably
consistent progress.

I've also added an optional paramete to change the chunk size. This was
partly to make testing it easier but also because we, or others who may
use this library may feel the need to change this value.

I've also changed the name of the function to be more descriptive.