Merge declaration files.
We were using dts-bundler to merge the two typescript declaration files
into one which is how the front end needs it.

dst-bundler is not maintained and more or less every dependency had a
CVE against it. Most of the other typescript bundlers haven't been
touched in years either.

All we need to do is merge the two files. This is quick and dirty but it
does work and it saves us having to install a new dependency to do the
job for us. I've tested this locally and it does work with the frontend.